+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Changelog
 
 ### Changes
 
+- **[Gradle]** Add missing experimental annotations to the Gradle plugin's analysis APIs. Sorry these were not meant to be stabilized yet!
 - Test Kotlin `2.4.0-RC2`.
 
 ### Contributors